pyproject.toml
[tool.poetry]
name = "pdc"
version = "0.1.0"
description = ""
authors = ["Carmelo San Giovanni <1@3bsd.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
faust-streaming = "^0.10.11"
aiohttp = "^3.8.4"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
virtualenvs.in-project = true
